Performance_agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langgraph>=0.2.0          
langchain>=0.3.0          
langchain-core>=0.3.0     
langchain-anthropic>=0.2.0 
langchain-openai>=0.2.0   
jinja2>=3.1.0             
jinja2>=3.1.0             
jinja2>=3.1.0             
pydantic>=2.5.0           

python-dotenv>=1.0.0

tiktoken>=0.7.0           
# anthropic>=0.25.0         


boto3>=1.34.0
aioboto3>=12.0.0          


pyyaml>=6.0               
rich>=13.0.0              

# Optional: Async HTTP
aiohttp>=3.9.0            # Async HTTP client

# Development
pytest>=7.4.0             
pytest-asyncio>=0.21.0    
